# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ultra ERP API Registry",
        "description": "Global modular API documentation for the Production-grade ERP system.",
        "contact": {
            "email": "admin@erp.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Primary API Gateway"
        }
    ],
    "paths": {
...
        "/api/v1/accounts": {
            "get": {
                "tags": [
                    "Accounting"
                ],
                "summary": "List all ledger accounts",
                "operationId": "f4d259d259913f9de3bffdbb1295870e",
                "parameters": [
                    {
                        "name": "active_only",
                        "in": "query",
                        "description": "Filter by active status",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by account type (asset, liability, equity, revenue, expense)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "current_page": 1,
                                        "data": [
                                            {
                                                "id": 1,
                                                "name": "Cash equivalent",
                                                "code": "1000",
                                                "type": "asset",
                                                "is_active": true
                                            }
                                        ],
                                        "total": 1
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Accounting"
                ],
                "summary": "Create new ledger account",
                "operationId": "ba8487268240c2e12a7077533f962bf9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "code",
                                    "type"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "asset",
                                            "liability",
                                            "equity",
                                            "revenue",
                                            "expense"
                                        ]
                                    },
                                    "branch_id": {
                                        "type": "integer"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Account created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Account created.",
                                    "data": {
                                        "id": 2,
                                        "name": "Accounts Receivable",
                                        "code": "1200",
                                        "type": "asset",
                                        "is_active": true
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/accounts/{account}": {
            "get": {
                "tags": [
                    "Accounting"
                ],
                "summary": "Get account details with balance",
                "operationId": "844ef57d6b9c5a3e8f5b4bd59617aaa8",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "Account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": {
                                        "id": 1,
                                        "name": "Cash equivalent",
                                        "code": "1000",
                                        "type": "asset",
                                        "is_active": true,
                                        "balance": 15000.5
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Accounting"
                ],
                "summary": "Update existing ledger account",
                "operationId": "2ea4198616890a11db3c87657d80eab6",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "Account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "code": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "asset",
                                            "liability",
                                            "equity",
                                            "revenue",
                                            "expense"
                                        ]
                                    },
                                    "branch_id": {
                                        "type": "integer"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Account updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Account updated.",
                                    "data": {
                                        "id": 1,
                                        "name": "Primary Cash",
                                        "code": "1000",
                                        "type": "asset",
                                        "is_active": true
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Account not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Accounting"
                ],
                "summary": "Delete ledger account",
                "operationId": "f2d467fee00c5fe4571f934b3f03a2dc",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "description": "Account ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Account deleted successfully"
                    },
                    "404": {
                        "description": "Account not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Enter token in 'Bearer {token}' format",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}